+++ src/Plugin.php
@@ -58,7 +58,7 @@ class Plugin extends BasePlugin
 
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '0.29.0';
+    public string $schemaVersion = '0.29.1';
 
     public static ?self $plugin = null;